requirements.txt
python-can>=4.2.2
pyserial>=3.5
git+https://github.com/brendan-w/python-OBD.git#egg=obd
customtkinter>=5.1.3
pillow>=9.5.0
pyusb>=1.2.1
